Bugfix: ArchiveFilesWorkPlan must always set metadata
Changed the predicate that controls archiving (which covers confirming
data matches checksums, upload, adding metadata) from:

RequiresArchiving

which will simply identify files by regex, to:

And(RequiresArchiving, Not(isArchived)),

which will identify files by regex and test them to see that they have
the exist in iRODS with right checksum and metadata. If not, all those
operations are reiterated (idempotently), doing any that are required.

Tweaked the logging descriptions for the workplans to be clearer.