Add compressed file handling.
Change compressed file suffix to .gz

Add IsCompressed().

HasChecksumFile now checks if a compressed version of the file has a checksum.

MakeIsArchived() now takes a mabyeCompressed bool, which if true will cause
the compressed version of the file (if it exists) to be archived instead of the
uncompressed version. (untested breaking change; fixed in subsequent commits)