Bugfix: DoProcessFiles deadlock
In DoProcessFiles, don't increment the WaitGroup until after the skip
test for files being worked on already. In its current state, the
WorkGroup is incremented and never decremeted on skips, causing the a
deadlock when shutting down.

Renamed doWork to makeWork It's a name more appropriate for what it
does.