Directory walker to return filepath.SkipDir in cancel
This change causes the directory walk function to always return
filepath.SkipDir once a operation cancel has been started. Previously
this function returned nil in these circumstances, allowing the
traversal to continue.

This change also brings the cancellation test to precede any action on
the current path so that it will not be passed to the work channel.