Checksum creation improvements.
Tee both uncompressed and compressed fastq through MD5, the former for
validating the original data, the latter to avoid traversing the file a
second time to create the checkum file for the compressed data.

Create checksum files in TMPDIR and them rename them, to avoid race
conditions with the directory walker.

Use encoding/hex to encode checksums, instead of rolling our own.