Fix test timeouts.
FindFilesInterval starts a number of goroutines sending to the paths
channel, which is unbuffered. When the operation is cancelled, some
may be still trying to write to the channel. As the test has been
cancelled at this point, nothing is consuming from the channel and
they block. This also blocks closing the errs channel.

In order to close and check the errs channel post-test, we need to
consume any remaining paths.