Improve handling of errors during processing.
Added an error processing goroutine in ProcessFiles. This constantly
drains the error channel to a log.

Added a noCancelMsg channel in ProcessFiles to allow the remaining
goroutine to exit, should processing complete without cancelling (it
shouldn't, but in case).

Added an error processing goroutine.

Refactored WatchFiles to forwards errors to the error channel, when
setting up, while detecting files and when completing.

Removed the size 1 buffer from some error channels where it is not
necessary (now that there is a goroutine draining the channel).