Add valet archive annotate command
The new "archive annotate" command ensures that metadata (primary
metadata at the moment) can be updated in the event of changes
e.g. sample swaps.

Currently all annotation comes from the MinKNOW report Markdown file
and is added to the collection containing that file in iRODS (not to
the file itself). This is to avoid adding metadata to each of the many
hundreds of files generated by the sequencing instrument.

This commit contains some organisation of how command line flags are
stored internally, so that they are separated by purpose and can be
re-used.