Remove run directories after a delay, if they are empty
This prevents an accumulation of empty directories on the instrument
after data have been transferred elsewhere. The run directories
contain the bulk of the empty directories, particularly if barcoding
has been used.

Directories are not removed if they or their descendants contain one or
more files.

A new command line option for the archive command allows the age
threshold of a deletable run directory to be specified, defaulting to
14 days, with a minimum of 60 seconds.